LayoutTests/svg/animations/animVal-basics-expected.txt

CONSOLE MESSAGE: line 3: ReferenceError: Can't find variable: createSVGTestCase
SVG 1.1 dynamic animation tests

Trivial animVal testcase, to see wheter we support it at all. Should result in a 200x200 rect and only PASS messages

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".
